Reduce the number of arguments in constructors
This change was a huge one, we reduced the number of arguments in the constructors of two classed from 9 and 10 down to 5.

Since bluetooth is used as a platform for several other components, we also had to fix the code in those components, additionaly we also fixed some tests for bluetooth itself.

Co-authored-by: latiif <latiif@mail.com>
Co-authored-by: oswalchlm <waoscar@student.chalmers.se>